+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [x] Context menu command to enable compiler for project
 - [x] Adornment to toggle compilation per solution
 - [x] Command to compile less files in all enabled projects in solution
+- [x] Make project/solution compiler defaults configurable
 - [ ] Compile only parent files and not all files
 
 Features that have a checkmark are complete and available for